Saved Changes

+++ calculator.py
@@ -1,5 +1,5 @@
 def main():
-    x= int(input("What is x?"))
+    x= float(input("What is x?"))
     print("x squared is", square(x))
 
 def square(n):
